+++ praw/models/reddit/mixins/__init__.py
@@ -14,6 +14,8 @@ from .votable import VotableMixin  # NOQA
 class ThingModerationMixin(object):
     """Provides moderation methods for Comments and Submissions."""
 
+    REMOVAL_MESSAGE_API = None
+
     def approve(self):
         """Approve a :class:`~.Comment` or :class:`~.Submission`.
 
@@ -111,6 +113,61 @@ class ThingModerationMixin(object):
         """
         data = {'id': self.thing.fullname, 'spam': bool(spam)}
         self.thing._reddit.post(API_PATH['remove'], data=data)
+
+    def send_removal_message(self, type,  # pylint: disable=redefined-builtin
+                             title=None, message=None):
+        """Send a removal message for a Comment or Submission.
+
+        Reddit adds human-readable information about the object to the message.
+
+        :param type: One of 'public', 'private', 'private_exposed'.
+            'public' leaves a stickied comment on the post.
+            'private' sends a Modmail message with hidden username.
+            'private_exposed' sends a Modmail message without hidden username.
+        :param title: The short reason given in the message.
+            (Ignored if type is 'public'.)
+        :param message: The body of the message.
+        """
+        # The API endpoint used to send removal messages is different
+        # for posts and comments, so the derived classes specify which one.
+        if self.REMOVAL_MESSAGE_API is None:
+            raise NotImplementedError('ThingModerationMixin must be extended.')
+        url = API_PATH[self.REMOVAL_MESSAGE_API]
+
+        # Only the first element of the item_id list is used.
+        data = {"item_id": [self.thing.fullname],
+                "message": message,
+                "title": title,
+                "type": str(type)}
+
+        # Use the core to make the request in order to send the data as
+        # JSON - this endpoint doesn't like URL encoding.
+        self.thing._reddit._core.request('POST', url, json=data)
+
+    def set_removal_reason(self, reason=None, mod_note="", other_things=None):
+        """Set a removal reason for a Comment or Submission.
+
+        Note that the removal message is not sent.  This can be done separately
+        using :meth:`~.send_removal_message`.
+
+        You can provide no reason if you only want to set mod_note.
+
+        :param reason: The :class:`~.RemovalReason` to use.
+        :param mod_note: A private note shown to moderators alongside the
+            reason.
+        :param other_things: When provided, set the same reason and mod_note on
+            this list of :class:`~.Comment` and/or :class:`~.Submission`
+            instances as part of a single request (default: None).
+        """
+        item_ids = [self.thing.fullname]
+        if other_things is not None:
+            item_ids += [t.fullname for t in other_things]
+        data = {'item_ids': item_ids,
+                'reason_id': None if reason is None else str(reason),
+                'mod_note': mod_note}
+
+        url = API_PATH['removal_reason_set']
+        self.thing._reddit._core.request('POST', url, json=data)
 
     def undistinguish(self):
         """Remove mod, admin, or special distinguishing on object.